It's very wierd.
When I comment out processReplacing of Distorsion, it will not crash.
But, if I remove comment it will crash.
And another but, if only one instance of plugin is hosted it will not
crash no matter how long(many times) processReplacing is called.

It will crash on load if I load second or third instance.

And also, if I load 10 instance of uncrashing plugin and Save and Exit,
if I reopen REAPER.exe(since im using reaper for testing) it will not
crash but if I load just one more plugin it will crash.

I can't seem to solve it..